Refund button view gone on item wise partial refund in TransactionDetails.

+++ app/src/main/java/com/pays/pos/ui/fragments/transactions/TransactionDetailsFragment.kt
@@ -1470,7 +1470,7 @@ class TransactionDetailsFragment : Fragment() {
                     * "total <= refundedAmount" This is commented so that a transaction can be refunded only once even if it was refunded partially.
                      */
 
-                    if (total <= refundedAmount /*refundedAmount.toDouble() > 0.0*/
+                    if (total.toDouble() <= refundedAmount.toDouble() /*refundedAmount.toDouble() > 0.0*/
                         || paymentDetailsResponse.data.order.payment_status == "Cancelled"
                     ) {
                         binding.tvIssueRefund.visibility = View.GONE
